requirements.txt
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
garth==0.5.17
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
seaborn==0.13.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
